requirements.txt
fastapi
uvicorn
openai
python-dotenv
agents
requests 